workloads/webhook-receiver/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-firestore==2.14.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
google-cloud-secret-manager==2.17.0
